+++ example/spec/parallel/testAPI/bulkOperationsSuccessSpec.js
@@ -4,7 +4,7 @@ const get = require('lodash/get');
 const pAll = require('p-all');
 
 const granules = require('@cumulus/api-client/granules');
-const { deleteAsyncOperation } = require('@cumulus/api-client/asyncOperations');
+const { deleteAsyncOperation, getAsyncOperation } = require('@cumulus/api-client/asyncOperations');
 const { deleteCollection } = require('@cumulus/api-client/collections');
 const { deleteExecution } = require('@cumulus/api-client/executions');
 const { deleteProvider } = require('@cumulus/api-client/providers');
@@ -16,7 +16,6 @@ const {
 } = require('@cumulus/aws-client/SQS');
 const { randomId } = require('@cumulus/common/test-utils');
 const {
-  api: apiTestUtils,
   getClusterArn,
   getExecutionInputObject,
 } = require('@cumulus/integration-tests');
@@ -175,11 +174,11 @@ describe('POST /granules/bulk', () => {
         console.log(`bulk operations async operation ID: ${postBulkOperationsBody.id}`);
 
         // Query the AsyncOperation API to get the task ARN
-        const getAsyncOperationResponse = await apiTestUtils.getAsyncOperation({
+        const asyncOperation = await getAsyncOperation({
           prefix,
-          id: postBulkOperationsBody.id,
+          asyncOperationId: postBulkOperationsBody.id,
         });
-        ({ taskArn } = JSON.parse(getAsyncOperationResponse.body));
+        ({ taskArn } = asyncOperation);
       } catch (error) {
         beforeAllFailed = true;
         console.log(error);
@@ -235,16 +234,11 @@ describe('POST /granules/bulk', () => {
     it('creates an AsyncOperation', async () => {
       if (beforeAllFailed) fail('beforeAll() failed');
       else {
-        const getAsyncOperationResponse = await apiTestUtils.getAsyncOperation({
+        const asyncOperation = await getAsyncOperation({
           prefix,
-          id: postBulkOperationsBody.id,
+          asyncOperationId: postBulkOperationsBody.id,
         });
-
-        expect(getAsyncOperationResponse.statusCode).toEqual(200);
-
-        const getAsyncOperationBody = JSON.parse(getAsyncOperationResponse.body);
-
-        expect(getAsyncOperationBody.id).toEqual(postBulkOperationsBody.id);
+        expect(asyncOperation.id).toEqual(postBulkOperationsBody.id);
       }
     });
 
@@ -272,26 +266,23 @@ describe('POST /granules/bulk', () => {
           }
         ).promise();
 
-        const getAsyncOperationResponse = await apiTestUtils.getAsyncOperation({
+        const asyncOperation = await getAsyncOperation({
           prefix,
-          id: postBulkOperationsBody.id,
+          asyncOperationId: postBulkOperationsBody.id,
         });
 
-        const getAsyncOperationBody = JSON.parse(getAsyncOperationResponse.body);
-
-        expect(getAsyncOperationResponse.statusCode).toEqual(200);
-        expect(getAsyncOperationBody.status).toEqual('SUCCEEDED');
+        expect(asyncOperation.status).toEqual('SUCCEEDED');
 
         let output;
         try {
-          output = JSON.parse(getAsyncOperationBody.output);
+          output = JSON.parse(asyncOperation.output);
         } catch (error) {
-          throw new SyntaxError(`getAsyncOperationBody.output is not valid JSON: ${getAsyncOperationBody.output}`);
+          throw new SyntaxError(`asyncOperation.output is not valid JSON: ${asyncOperation.output}`);
         }
 
         await getGranuleWithStatus({
           prefix,
-          granuleId: JSON.parse(getAsyncOperationBody.output)[0],
+          granuleId: JSON.parse(asyncOperation.output)[0],
           status: 'running',
           timeout: 120,
           updatedAt: ingestedGranule.updatedAt,