+++ packages/api/tests/lib/writeRecords/test-write-execution.js
@@ -17,8 +17,9 @@ const Execution = require('../../../models/executions');
 const {
   buildExecutionRecord,
   shouldWriteExecutionToPostgres,
-  writeExecution,
-} = require('../../../lambdas/sf-event-sqs-to-db-records/write-execution');
+  writeExecutionRecordFromMessage,
+  writeExecutionRecordFromApi,
+} = require('../../../lib/writeRecords/write-execution');
 
 test.before(async (t) => {
   process.env.ExecutionsTable = cryptoRandomString({ length: 10 });
@@ -222,7 +223,7 @@ test('buildExecutionRecord returns record with error', (t) => {
   t.deepEqual(record.error, exception);
 });
 
-test('writeExecution() saves execution to Dynamo and RDS if write to RDS is enabled', async (t) => {
+test('writeExecutionRecordFromMessage() saves execution to Dynamo and RDS if write to RDS is enabled', async (t) => {
   const {
     cumulusMessage,
     knex,
@@ -231,13 +232,13 @@ test('writeExecution() saves execution to Dynamo and RDS if write to RDS is enab
     executionPgModel,
   } = t.context;
 
-  await writeExecution({ cumulusMessage, knex });
+  await writeExecutionRecordFromMessage({ cumulusMessage, knex });
 
   t.true(await executionModel.exists({ arn: executionArn }));
   t.true(await executionPgModel.exists(knex, { arn: executionArn }));
 });
 
-test('writeExecution() saves execution to Dynamo and RDS with same timestamps', async (t) => {
+test('writeExecutionRecordFromMessage() saves execution to Dynamo and RDS with same timestamps', async (t) => {
   const {
     cumulusMessage,
     knex,
@@ -246,7 +247,7 @@ test('writeExecution() saves execution to Dynamo and RDS with same timestamps', 
     executionPgModel,
   } = t.context;
 
-  await writeExecution({ cumulusMessage, knex });
+  await writeExecutionRecordFromMessage({ cumulusMessage, knex });
 
   const dynamoRecord = await executionModel.get({ arn: executionArn });
   const pgRecord = await executionPgModel.get(knex, { arn: executionArn });
@@ -254,7 +255,7 @@ test('writeExecution() saves execution to Dynamo and RDS with same timestamps', 
   t.is(pgRecord.updated_at.getTime(), dynamoRecord.updatedAt);
 });
 
-test.serial('writeExecution() does not persist records to Dynamo or RDS if Dynamo write fails', async (t) => {
+test.serial('writeExecutionRecordFromMessage() does not persist records to Dynamo or RDS if Dynamo write fails', async (t) => {
   const {
     cumulusMessage,
     knex,
@@ -264,13 +265,13 @@ test.serial('writeExecution() does not persist records to Dynamo or RDS if Dynam
   } = t.context;
 
   const fakeExecutionModel = {
-    storeExecutionFromCumulusMessage: () => {
+    storeExecutionRecord: () => {
       throw new Error('execution Dynamo error');
     },
   };
 
   await t.throwsAsync(
-    writeExecution({
+    writeExecutionRecordFromMessage({
       cumulusMessage,
       knex,
       executionModel: fakeExecutionModel,
@@ -281,7 +282,7 @@ test.serial('writeExecution() does not persist records to Dynamo or RDS if Dynam
   t.false(await executionPgModel.exists(knex, { arn: executionArn }));
 });
 
-test.serial('writeExecution() does not persist records to Dynamo or RDS if RDS write fails', async (t) => {
+test.serial('writeExecutionRecordFromMessage() does not persist records to Dynamo or RDS if RDS write fails', async (t) => {
   const {
     cumulusMessage,
     knex,
@@ -302,14 +303,14 @@ test.serial('writeExecution() does not persist records to Dynamo or RDS if RDS w
   t.teardown(() => trxStub.restore());
 
   await t.throwsAsync(
-    writeExecution({ cumulusMessage, knex }),
+    writeExecutionRecordFromMessage({ cumulusMessage, knex }),
     { message: 'execution RDS error' }
   );
   t.false(await executionModel.exists({ arn: executionArn }));
   t.false(await executionPgModel.exists(knex, { arn: executionArn }));
 });
 
-test.serial('writeExecution() correctly sets both original_payload and final_payload in postgres when execution records are run in sequence', async (t) => {
+test.serial('writeExecutionRecordFromMessage() correctly sets both original_payload and final_payload in postgres when execution records are run in sequence', async (t) => {
   const {
     cumulusMessage,
     knex,
@@ -322,13 +323,35 @@ test.serial('writeExecution() correctly sets both original_payload and final_pay
 
   cumulusMessage.meta.status = 'running';
   cumulusMessage.payload = originalPayload;
-  await writeExecution({ cumulusMessage, knex });
+  await writeExecutionRecordFromMessage({ cumulusMessage, knex });
 
   cumulusMessage.meta.status = 'completed';
   cumulusMessage.payload = finalPayload;
-  await writeExecution({ cumulusMessage, knex });
+  await writeExecutionRecordFromMessage({ cumulusMessage, knex });
 
   const pgRecord = await executionPgModel.get(knex, { arn: executionArn });
   t.deepEqual(pgRecord.original_payload, originalPayload);
   t.deepEqual(pgRecord.final_payload, finalPayload);
+});
+
+test('writeExecutionRecordFromApi() saves execution to Dynamo and RDS with same timestamps', async (t) => {
+  const {
+    cumulusMessage,
+    knex,
+    executionModel,
+    executionArn,
+    executionPgModel,
+  } = t.context;
+
+  const apiRecord = Execution.generateRecord(cumulusMessage);
+  await writeExecutionRecordFromApi({
+    record: apiRecord,
+    knex,
+    executionModel,
+  });
+
+  const dynamoRecord = await executionModel.get({ arn: executionArn });
+  const pgRecord = await executionPgModel.get(knex, { arn: executionArn });
+  t.is(pgRecord.created_at.getTime(), dynamoRecord.createdAt);
+  t.is(pgRecord.updated_at.getTime(), dynamoRecord.updatedAt);
 });